# src/main/resources/openapi/api.yaml
...
openapi: 3.0.1
info:
  title: Fast Food Catalog API
  version: v0
paths:
...
  /catalog:
    get:
      description: Busca paginada de produtos ativos por categoria
      tags:
        - Catalogo
      operationId: getByCategory
      parameters:
        - name: category
          description: Categoria do produto
          in: query
          required: true
          schema:
            type: string
            enum:
              - SANDWICH
              - DRINK
              - SIDE_DISH
              - DESSERT
        - name: page
          description: Número da página
          in: query
          required: true
          schema:
            minimum: 1
            type: integer
            format: int32
        - name: size
          description: Quantidade de elementos por página
          in: query
          required: true
          schema:
            maximum: 50
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageResultProductView'
  '/catalog/{id}':
    get:
      description: Busca um produto pelo id
      tags:
        - Catalogo
      operationId: getById
      parameters:
        - name: id
          description: ID do produto
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductView'
components:
  schemas:
...
    ProductView:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        images:
          uniqueItems: true
          type: array
          items:
            type: string
        price:
          type: number
        category:
          type: string
          enum:
            - SANDWICH
            - DRINK
            - SIDE_DISH
            - DESSERT
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
    PageResultView:
      type: object
      properties:
        pageNumber:
          description: Número da página
          type: integer
          format: int32
        pageSize:
          description: Quantidade de elementos por página
          type: integer
          format: int32
        totalElements:
          description: Quantidade total de elementos
          type: integer
          format: int64
        totalPages:
          description: Quantidade total de páginas
          type: integer
          format: int32
    PageResultProductView:
      allOf:
        - $ref: '#/components/schemas/PageResultView'
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProductView'